+++ src/FieldLayout/LayoutElements/Users/EmailField.php
@@ -70,7 +70,7 @@ class EmailField extends TextField implements ImportableFieldLayoutElementInterf
     }
 
     #[Override]
-    protected function warning(?ElementInterface $element = null, bool $static = false): ?string
+    protected function warningText(?ElementInterface $element = null, bool $static = false): ?string
     {
         /** @var User $element */
         if (